Fix real time order table component
